+++ app/doc/open_api/v1/invoices_doc.rb
@@ -26,8 +26,10 @@ class OpenAPI::V1::InvoicesDoc < OpenAPI::V1::BaseDoc
             "invoiced_id": 3257,
             "user_id": 211,
             "invoiced_type": "Reservation",
-            "stp_invoice_id": "in_187DLE4zBvgjueAZ6L7SyQlU",
-            "stp_payment_intent_id": "pi_1Dat4P2eZvKYlo2C3MxszwQp",
+            "payment_gateway_object": {
+              id: "in_187DLE4zBvgjueAZ6L7SyQlU",
+              type: "Stripe::Invoice"
+            },
             "reference": "1605017/VL",
             "total": 1000,
             "type": null,
@@ -42,8 +44,10 @@ class OpenAPI::V1::InvoicesDoc < OpenAPI::V1::BaseDoc
             "invoiced_id": 3229,
             "user_id": 211,
             "invoiced_type": "Reservation",
-            "stp_invoice_id": "in_185Hmt4zBvgjueAZl5lio1pK",
-            "stp_payment_intent_id": "pi_2Dat4P2eYbKYlo2C3MxszwQp",
+            "payment_gateway_object": {
+              id: "pi_2Dat4P2eYbKYlo2C3MxszwQp",
+              type: "Stripe::PaymentIntent"
+            },
             "reference": "1604176/VL",
             "total": 2000,
             "type": null,
@@ -58,8 +62,10 @@ class OpenAPI::V1::InvoicesDoc < OpenAPI::V1::BaseDoc
             "invoiced_id": 3218,
             "user_id": 211,
             "invoiced_type": "Reservation",
-            "stp_invoice_id": "in_184oNK4zBvgjueAZJdOxHJjT",
-            "stp_payment_intent_id": "pi_1Pub4P2eZvKYlo2C3MxszwQm",
+            "payment_gateway_object": {
+              id: "ba15dc9d8f3e0fa17bf527466",
+              type: "PayZen::Order"
+            },
             "reference": "1604166/VL",
             "total": 2000,
             "type": null,