+++ app/models/payment_gateway_object.rb
@@ -6,6 +6,10 @@ require 'payment/item_builder'
 class PaymentGatewayObject < ApplicationRecord
   belongs_to :item, polymorphic: true
   belongs_to :invoice, foreign_type: 'Invoice', foreign_key: 'item_id'
+  belongs_to :invoice_item, foreign_type: 'InvoiceItem', foreign_key: 'item_id'
+  belongs_to :subscription, foreign_type: 'Subscription', foreign_key: 'item_id'
+  belongs_to :payment_schedule, foreign_type: 'PaymentSchedule', foreign_key: 'item_id'
+  belongs_to :payment_schedule_item, foreign_type: 'PaymentScheduleItem', foreign_key: 'item_id'
 
   def gateway_object
     item = Payment::ItemBuilder.build(gateway_object_type)