+++ app/workers/payment_schedule_item_worker.rb
@@ -18,15 +18,15 @@ class PaymentScheduleItemWorker
 
   def check_item(psi)
     # the following depends on the payment method (stripe/check)
-    if psi.payment_schedule.payment_method == 'stripe'
+    if psi.payment_schedule.payment_method == 'card'
       ### Stripe
       stripe_key = Setting.get('stripe_secret_key')
       stp_subscription = Stripe::Subscription.retrieve(psi.payment_schedule.stp_subscription_id, api_key: stripe_key)
       stp_invoice = Stripe::Invoice.retrieve(stp_subscription.latest_invoice, api_key: stripe_key)
       if stp_invoice.status == 'paid'
         ##### Stripe / Successfully paid
-        PaymentScheduleService.new.generate_invoice(psi, payment_method: 'stripe', payment_id: stp_invoice.payment_intent)
-        psi.update_attributes(state: 'paid', payment_method: 'stripe', stp_invoice_id: stp_invoice.id)
+        PaymentScheduleService.new.generate_invoice(psi, payment_method: 'card', payment_id: stp_invoice.payment_intent, payment_type: 'Stripe::PaymentIntent') # FIXME
+        psi.update_attributes(state: 'paid', payment_method: 'card', stp_invoice_id: stp_invoice.id)
       elsif stp_subscription.status == 'past_due' || stp_invoice.status == 'open'
         ##### Stripe / Payment error
         if psi.state == 'new'